Allow parameters to be good as well as bad
This requires some additional sophistication in the algorithm. It
was hard to balance the old naive prior with the concept of good
elements without either some very ad hoc numbers or weird special
casing.

Fortunately there's an elegant solution: We switch to a prior which
is empirically derived from the data. This allows us to adapt to
differing ratios of good and bad things in the population.

This is not currently hooked into anything but will be one of the
major mechanisms required for coverage driven example discovery.