Reintroduce the notion of complexity
Complexity is an additional feature that allows Hypothesis to guide
the searching. It allows you to say which of two examples you prefer.

In straightforward falsification this turns out to be not very useful
when compared to the greedy search. You're better off just stopping as
soon as you find an example and minimizing it directly through simplify.

In interestingness detection however this is not the case: We will have
a large number of examples indexed by reasons they're interesting and
may discover new ones at any given time. This means that we need to
update our existing table of examples if the new example we found for
this happens to be better at the old one. This is why having a complexity
metric is useful.