A giant renaming
Concept: The examples that get produced may not be the examples that
get passed to the function being tested. Previously this was just
copying but for more sophisticated examples (i.e. the Django testing)
it may be something wildly different.

Thus instead of having values which we produce, we produce templates
which we then reify into values. For most things reifying will just
be copying, but the intent is that for e.g. Django tests the template
will be "Here is a sequence of database inserts you can perform" and
the reification will then actually perform those inserts.

99% of strategies will not have to care about this even slightly and
can just rename their produce method to produce_template. The
remaining 1% will appreciate this greatly.