Better alphabet selection
A lot of people are trying to get strings in a particular region
by just using assume. This works very badly because Hypothesis was
generating really high unicode ranges.

On the one hand this isn't a good way to generate strings, on the
other hand Hypothesis *should* be trying to generate more ascii
characters that it currently is.

This basically changes the alphabet selection algorithm so that
instead of generating over the whole range it's strongly biased
towards the ascii range.

The selection algorithm is totally ad hoc, but satisfies the new
statistical tests as to what probabilities it should generate:

1. Strings with all ascii
2. Long strings with no ascii
3. Strings with at least two ascii characters