Rebalance ints towards large values
This is a surprisingly big commit, fixing a bunch of interlocking issues.

Some bugs can only be triggered by producing integers which are too large
to fit into a machine word of various sizes. Lets make sure we can generate
those on a regular basis.

This then prompted some fixing of integer ranges which would have resulted
in incredibly bad performance when used with large ranges.

This then prompted some interesting new experiments with cloning and
simplifying lists of examples.