Better handling of complexity in database
We're potentially putting much larger examples in the DB now. This
is a shame.

We cut off at half timeout and make sure to iterate from simplest
to most complex. This means that we should gradually improve the
quality of what's in the database by giving us time to simplify.