Rebuild flatmap on top of morphers
Flatmap is now just a map over a pair of the source and a morpher.

This should still have the desired effect because the second argument
of the pair can always be simplified. If anything this should
significantly improve the quality of the flatmap simplification because
it handles the case where the underlying strategy changes much more
robustly.