Fold in lessons learned about shrinking
The idea is that warmup is entirely the wrong approach. What we
want to do is bet on the multi-pass simplification working (because
it usually does) and then reboot every now and then, gradually
degrading to the greedy algorithm. This lets us continue to exploit
structure much more readily but is able to meet the greedy algorithm
halfway on less well structured examples.

This cuts about 30% run time off the example quality benchmark.